Development

+++ backend/controllers/messageController.js
@@ -5,10 +5,9 @@ const {StatusCodes} = require('http-status-codes');
 const Message = require('../models/message');
 const ResponseController = require('./responseController');
 const UserController = require("./userController");
+const {BadRequest} = require("../errors");
 
 class MessageController {
-
-    //Todo: all return statements coming from the database will be checked for whether they returned error or not
 
     static #message=new Message();
 
@@ -16,8 +15,12 @@ class MessageController {
     static async createMessage(request, response, next) {
         try {
             const senderId=UserController.fetchUserIdFromRequest(request);
-            const messageBody=request.body; // Todo: message body will be checked
-            const newMessage=await MessageController.#message.getModel().create({...messageBody,senderId});
+            const messageBody={...request.body,senderId};
+            if (!MessageController.#message.areRequiredDefinitionKeysMatched(messageBody)) {
+                throw new BadRequest('Message body is not valid!');
+            }
+            const newMessage=await MessageController.#message.getModel().create(messageBody);
+            if (!newMessage) {throw new BadRequest('Message could not be created!')};
             const responseData = ResponseController.createResponseData(newMessage);
             return response.status(StatusCodes.OK).json(responseData);
         } catch (error) {
@@ -36,7 +39,8 @@ class MessageController {
                 .limit(limit)
                 .skip(skipCount)
                 .sort({'createdAt' : -1})
-            const totalMessages = await MessageController.#message.getModel().estimatedDocumentCount();
+            if (!messagesOfConversation) {throw new BadRequest('messages of conversation is not found!')}
+            const totalMessages = await MessageController.#message.getModel().countDocuments({conversationId});
             const totalPages = Math.ceil(totalMessages / limit);
             const paginationInfo = {
                 messagesOfConversation,
@@ -60,6 +64,7 @@ class MessageController {
                 .find({conversationId, createdAt : {$lte : new Date(Number(nextCursorTimeStamp))}})
                 .limit(Number(limit) + 1)
                 .sort({createdAt : -1})
+            if (!messagesOfConversation) {throw new BadRequest('messages of conversation is not found!')}
             nextCursorTimeStamp=-1;
             const isFinished=messagesOfConversation.length === 1;
             if (!isFinished) {
@@ -85,6 +90,7 @@ class MessageController {
         try {
             const conversationId= request.params.conversationId;
             const messagesOfConversation = await MessageController.#message.getModel().find({conversationId});
+            if (!messagesOfConversation) {throw new BadRequest('messages of conversation is not found!')}
             const responseData = ResponseController.createResponseData(messagesOfConversation);
             return response.status(StatusCodes.OK).json(responseData);
         } catch (error) {
@@ -96,8 +102,10 @@ class MessageController {
         try {
             const senderId=UserController.fetchUserIdFromRequest(request);
             const messageId=request.params.messageId;
+            if (!messageId) {throw new BadRequest('message id is not found!')}
             const removedMessage=await MessageController.#message.getModel()
                 .findOneAndDelete({senderId, _id : messageId});
+            if (!removedMessage) {throw new BadRequest('message could not be removed!')}
             const responseData=ResponseController.createResponseData(removedMessage);
             return response.status(StatusCodes.OK).json(responseData);
         } catch (error) {
